refactor: moved fetching logic
